fy25/pyproject.toml
[project]
name = "northwestern-agentic-ai"
version = "0.1.0"
description = "Teaching repository for Northwestern University's MS AI Foundry program - Building Agentic AI Systems"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.14"
authors = [
    { name = "Northwestern MS AI Program" }
]
keywords = ["ai", "agents", "azure", "foundry", "mcp", "teaching"]

dependencies = [
    # Azure AI and Core SDKs - Microsoft Agent Framework
    # Note: agent-framework uses pre-release versions (beta)
    "agent-framework>=1.0.0b260130",
    "agent-framework-azure-ai>=1.0.0b260130",
    "azure-ai-projects>=2.0.0b2",
    "azure-ai-agents>=1.2.0b5",
    "azure-ai-evaluation>=1.0.0",
    "azure-ai-contentsafety>=1.0.0",
    "azure-identity>=1.25.0",
    "azure-core>=1.38.0",
    # Azure Functions and Logic Apps
    "azure-functions>=1.24.0",
    "azure-mgmt-logic>=10.0.0",
    "azure-mgmt-web>=7.3.0",
    # AI and ML
    "openai>=1.99.0",
    "python-dotenv>=1.2.0",
    # UI for Notebook Chat Interface
    "gradio>=4.0.0",
    # Utilities
    "pydantic>=2.10.0",
    "requests>=2.32.0",
    "aiohttp>=3.11.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    # Development and Testing
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    # Code Quality
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pyright>=1.1.0",
    "types-requests>=2.32.0",
    # Task Runner
    "poethepoet>=0.31.0",
    # Pre-commit
    "pre-commit>=4.0.0",
]

[project.urls]
repository = "https://github.com/northwestern/northwestern-fy26-msai-foundry-agentic-ai"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ============================================================
# Poe the Poet - Task Runner
# ============================================================

[tool.poe]
poetry_command = ""

[tool.poe.tasks]
# Setup and Installation
[tool.poe.tasks.setup]
help = "Set up the development environment with venv, dependencies, and pre-commit hooks"
sequence = ["venv", "install", "pre-commit-install"]

[tool.poe.tasks.venv]
help = "Create a virtual environment with specified Python version"
cmd = "uv venv"

[tool.poe.tasks.install]
help = "Install all dependencies including dev extras"
cmd = "uv sync --all-extras --dev"

[tool.poe.tasks.pre-commit-install]
help = "Install pre-commit hooks"
cmd = "pre-commit install"

# Code Quality and Formatting
[tool.poe.tasks.fmt]
help = "Format code using ruff"
sequence = [
    { cmd = "ruff format src/ tests/ notebooks/" },
]

[tool.poe.tasks.lint]
help = "Run linting checks and fix issues"
sequence = [
    { cmd = "ruff check --fix src/ tests/" },
]

[tool.poe.tasks.pyright]
help = "Run Pyright type checking"
cmd = "pyright src/"

[tool.poe.tasks.mypy]
help = "Run MyPy type checking"
cmd = "mypy src/"

# Testing
[tool.poe.tasks.test]
help = "Run unit tests with coverage"
cmd = "pytest tests/ --cov=src --cov-report=term-missing --cov-report=html"

[tool.poe.tasks.test-unit]
help = "Run only unit tests"
cmd = "pytest tests/ -m unit"

[tool.poe.tasks.test-integration]
help = "Run integration tests (requires RUN_INTEGRATION_TESTS=true)"
cmd = "pytest tests/ -m integration"

# Comprehensive Checks
[tool.poe.tasks.check]
help = "Run all quality checks (format, lint, pyright, mypy, test)"
sequence = ["fmt", "lint", "pyright", "test"]

# Cleanup
[tool.poe.tasks.clean]
help = "Clean up cache and build files"
cmd = """
rm -rf .venv __pycache__ .pytest_cache .ruff_cache .mypy_cache htmlcov .coverage dist/ *.egg-info/
find . -type d -name "__pycache__" -exec rm -rf {} + 2>/dev/null || true
find . -type f -name "*.pyc" -delete 2>/dev/null || true
"""

# ============================================================
# Tool Configuration
# ============================================================

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require external services)",
    "slow: Slow running tests",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["**/__pycache__", ".venv"]
